pyproject.toml
[project]
name = "MADS_thesis_equivariance"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Falco van den Hoogen", email = "falco.vandenhoogen@gmail.com"},
]
dependencies = ["torch>=2.8.0", "torchvision>=0.23.0", "e2cnn>=0.2.3", "matplotlib>=3.10.3", "pillow>=11.3.0", "loguru>=0.7.3", "ray[tune]>=2.47.1", "scikit-learn>=1.7.0", "sahi>=0.11.36", "pycocotools>=2.0.10", "numpy>=2.3.3"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
